+++ src/module/Normalizer/src/Normalizer/Adapter/AbstractAdapter.php
@@ -25,9 +25,11 @@ namespace Normalizer\Adapter;
 use Normalizer\Entity\Normalized;
 use DateTime;
 use Normalizer\Exception\RuntimeException;
+use Zend\I18n\Translator\TranslatorAwareTrait;
 
 abstract class AbstractAdapter implements AdapterInterface
 {
+    use TranslatorAwareTrait;
     protected $object;
 
     public function getObject()
@@ -60,11 +62,13 @@ abstract class AbstractAdapter implements AdapterInterface
             'routeParams' => $this->getRouteParams(),
             'id'          => $this->getId(),
             'metadata'    => [
-                'creationDate' => $this->getCreationDate() ? $this->getCreationDate() : new DateTime(),
-                'description'  => $this->getDescription(),
-                'keywords'     => $this->getKeywords(),
-                'lastModified' => $this->getLastModified() ? $this->getLastModified() : new DateTime(),
-                'robots'       => $this->isTrashed() ? 'noindex' : 'all',
+                'title'            => $this->getHeadTitle(),
+                'creationDate'     => $this->getCreationDate() ? $this->getCreationDate() : new DateTime(),
+                'description'      => $this->getDescription(),
+                'metaDescription'  => $this->getMetaDescription(),
+                'keywords'         => $this->getKeywords(),
+                'lastModified'     => $this->getLastModified() ? $this->getLastModified() : new DateTime(),
+                'robots'           => $this->isTrashed() ? 'noindex' : 'all',
             ],
         ]);
 
@@ -87,6 +91,14 @@ abstract class AbstractAdapter implements AdapterInterface
     protected function getDescription()
     {
         return $this->getContent();
+    }
+
+    /**
+     * @return string
+     */
+    protected function getMetaDescription()
+    {
+        return $this->getDescription();
     }
 
     /**
@@ -136,4 +148,12 @@ abstract class AbstractAdapter implements AdapterInterface
      * @return boolean
      */
     abstract protected function isTrashed();
+
+    /**
+     * @return string
+     */
+    protected function getHeadTitle()
+    {
+        return $this->getTitle();
+    }
 }