+++ src/module/Normalizer/src/Normalizer/Factory/NormalizeHelperFactory.php
@@ -38,7 +38,10 @@ class NormalizeHelperFactory implements FactoryInterface
     {
         $normalize  = new Normalize();
         $normalizer = $serviceLocator->getServiceLocator()->get('Normalizer\Normalizer');
+        $instanceManager = $serviceLocator->getServiceLocator()->get('Instance\Manager\InstanceManager');
+
         $normalize->setNormalizer($normalizer);
+        $normalize->setInstanceManager($instanceManager);
 
         return $normalize;
     }