+++ src/module/Normalizer/src/Normalizer/View/Helper/Normalize.php
@@ -23,14 +23,19 @@
 namespace Normalizer\View\Helper;
 
 use Common\Filter\PreviewFilter;
+use Instance\Manager\InstanceManagerAwareTrait;
+use Markdown\Service\OryRenderService;
 use Markdown\View\Helper\MarkdownHelper;
+use Markdown\View\Helper\OryFormatHelper;
 use Normalizer\NormalizerAwareTrait;
+use Ui\View\Helper\Brand;
 use Zend\View\Helper\AbstractHelper;
 use Zend\View\Helper\HeadMeta;
 use Zend\View\Helper\HeadTitle;
 
 class Normalize extends AbstractHelper
 {
+    use InstanceManagerAwareTrait;
     use NormalizerAwareTrait;
 
     public function __invoke($object = null)
@@ -43,41 +48,149 @@ class Normalize extends AbstractHelper
 
     public function headMeta($object)
     {
-        /* @var $meta HeadMeta */
-        $meta       = $this->getView()->plugin('headMeta');
+        $meta = $this->getView()->plugin('headMeta');
         $normalized = $this->normalize($object);
-        $title      = $normalized->getTitle();
-        $type       = $normalized->getType();
-        $metadata   = $normalized->getMetadata();
-        $keywords   = $metadata->getKeywords();
-        $robots     = $metadata->getRobots();
-        $preview    = $this->toPreview($object);
-        $image      = $this->getMetaImage($object);
-        $meta->setProperty('og:title', $title);
-        $meta->setProperty('og:image', $image);
+
+        $type = $normalized->getType();
+        $metadata = $normalized->getMetadata();
+
+        $keywords = $metadata->getKeywords();
+        $robots = $metadata->getRobots();
+
         $meta->appendName('content_type', $type);
-        $meta->appendName('description', $preview);
+        $meta->appendName('description', $this->getMetaDescription($object));
         $meta->appendName('keywords', implode(', ', $keywords));
         $meta->appendName('robots', $robots);
 
+        $this->appendOpenSearchMeta($object);
+        $this->appendOpenGraphMeta($object);
+        $this->appendFacebookMeta($object);
+
         return $this;
     }
 
-    public function getMetaImage($object)
+    private function appendOpenSearchMeta($object)
     {
-        //TODO: Change path depending ob subject
-        return 'https://de.serlo.org/assets/images/meta_serlo.jpg';
+        $lang = $this->getSubdomain();
+
+        if ($lang === 'de' || $lang === 'en') {
+            $this->getView()->headLink(
+                [
+                    'rel' => 'search',
+                    'href' => '/opensearch.' . $lang . '.xml',
+                    'title' => 'Serlo (' . $lang . ')',
+                    'type' => 'application/opensearchdescription+xml',
+                ]
+            );
+        }
     }
 
-    public function headTitle($object)
+    private function appendOpenGraphMeta($object)
+    {
+        $meta = $this->getMeta();
+
+        $meta->setProperty('og:title', $this->normalize($object)->getTitle());
+        $meta->setProperty('og:type', 'website');
+        $meta->setProperty('og:image', $this->getMetaImage($object));
+        $meta->setProperty('og:description', $this->getMetaDescription($object));
+        $meta->setProperty('og:site_name', $this->getView()->brand()->getBrand(true));
+    }
+
+    private function appendFacebookMeta($object)
+    {
+        $meta = $this->getMeta();
+
+        $meta->setProperty('fb:pages', '155020041197918');
+        $meta->setProperty('fb:profile_id', '155020041197918');
+    }
+
+
+    private function getMetaDescription($object)
+    {
+        $metadata = $this->normalize($object)->getMetadata();
+        return $this->renderPreview($metadata->getMetaDescription());
+    }
+
+    private function getMetaImage($object)
+    {
+        $fileName = 'serlo.jpg';
+
+        $subject = trim(strtolower(strip_tags(
+            $this->getView()->navigation('default_navigation')
+                ->menu()
+                ->setPartial('layout/navigation/partial/active-subject')
+                ->setOnlyActiveBranch(true)
+                ->setMinDepth(0)
+                ->setMaxDepth(0)
+                ->render()
+        )));
+
+        switch ($subject) {
+            case 'mathematik':
+                $fileName = 'de/mathematik.jpg';
+                break;
+
+            case 'angewandte nachhaltigkeit':
+                $fileName = 'de/angewandte-nachhaltigkeit.jpg';
+                break;
+
+            case 'biologie':
+                $fileName = 'de/biologie.jpg';
+                break;
+        }
+
+        return 'https://assets.serlo.org/meta/' . $fileName;
+    }
+
+    private function getSubdomain()
+    {
+        return $this->getInstanceManager()->getInstanceFromRequest()->getSubdomain();
+    }
+
+    /**
+     * @return HeadMeta
+     */
+    private function getMeta()
+    {
+        return $this->getView()->plugin('headMeta');
+    }
+
+    public function headTitle($object = null)
     {
         /* @var $headTitle HeadTitle */
         $headTitle  = $this->getView()->plugin('headTitle');
-        $normalized = $this->normalize($object);
-        $title      = $normalized->getTitle();
+        if ($object == null) {
+            /** @var Brand $brand */
+            $brand  = $this->getView()->brand();
+            $title = $brand->getBrand(true) . " – " . $brand->getSlogan();
+        } elseif (is_string($object)) {
+            $title = $this->appendBrand($object);
+        } else {
+            $normalized = $this->normalize($object);
+            $title = $this->appendBrand($normalized->getMetadata()->getTitle());
+        }
         $headTitle($title);
-
         return $this;
+    }
+
+    private function appendBrand($title)
+    {
+        /** @var Brand $brand */
+        $brand  = $this->getView()->brand();
+        $maxStringLen = 65;
+
+        //add "– lernen mit Serlo"
+        $titlePostfix = ' – ' . $brand->getHeadTitle(true);
+        if (strlen($title) < ($maxStringLen-strlen($titlePostfix))) {
+            return $title . $titlePostfix;
+        }
+
+        $titlePostfixFallback = ' – ' . $brand->getBrand(true);
+        if (strlen($title) < ($maxStringLen-strlen($titlePostfixFallback))) {
+            return $title . $titlePostfixFallback;
+        }
+
+        return $title;
     }
 
     public function possible($object)
@@ -114,12 +227,20 @@ class Normalize extends AbstractHelper
 
     public function toPreview($object)
     {
-        /* @var $markdown MarkdownHelper */
-        $markdown   = $this->getView()->plugin('markdown');
         $normalized = $this->normalize($object);
+        $content    = $normalized->getMetadata()->getDescription();
+        return $this->renderPreview($content);
+    }
+
+    private function renderPreview($string)
+    {
+        $isOryEditorFormat = $this->getView()->plugin('isOryEditorFormat');
+        /** @var MarkdownHelper $renderer */
+        $renderer = $isOryEditorFormat($string)
+            ? $this->getView()->plugin('oryRenderer')
+            : $this->getView()->plugin('markdown');
+        $content =  $renderer->toHtml($string);
         $filter     = new PreviewFilter(152);
-        $content    = $normalized->getContent();
-        $content    = $markdown->toHtml($content);
         $preview    = $filter->filter($content);
         return $preview;
     }